Started writing new tests, tests for lexer() and parse() not yet implemented
